:hammer: make TokenInputRow expect/actual so mobile can provide a flicker-free input
